MVC, REST done (back-end)
