+++ src/main/java/com/proj/Matjalal/domain/article/controller/ApiV1ArticleController.java
@@ -1,10 +1,9 @@
 package com.proj.Matjalal.domain.article.controller;
 
-
 import com.proj.Matjalal.domain.article.entity.Article;
 import com.proj.Matjalal.domain.article.service.ArticleService;
+import com.proj.Matjalal.domain.ingredient.entity.Ingredient;
 import com.proj.Matjalal.global.RsData.RsData;
-import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
@@ -22,123 +21,116 @@ import java.util.Optional;
 public class ApiV1ArticleController {
     private final ArticleService articleService;
 
-    @AllArgsConstructor
+
+    //다건 조회 DTO
     @Getter
+    @AllArgsConstructor
     public static class ArticlesResponse {
-        private final List<Article> articles;
+        private final List<Article> Articles;
     }
 
+    //다건 조회
     @GetMapping("")
-    public RsData<ArticlesResponse> getArticles() {
-        List<Article> articles = this.articleService.getList();
+    private RsData<ArticlesResponse> getArticles() {
+        List<Article> Articles = this.articleService.getAll();
 
-
-        return RsData.of("S-1", "성공", new ArticlesResponse(articles));
+        return RsData.of("S-1", "성공", new ArticlesResponse(Articles));
     }
 
-    @AllArgsConstructor
+    //단건 조회 DTO
     @Getter
+    @AllArgsConstructor
     public static class ArticleResponse {
-        private final Article article;
+        private final Article Article;
     }
 
+    //단건 조회
     @GetMapping("/{id}")
-    public RsData<ArticleResponse> getArticle(@PathVariable("id") Long id) {
+    private RsData<ArticleResponse> getArticle(@PathVariable(value = "id") Long id) {
         return this.articleService.getArticle(id).map(article -> RsData.of(
-                "S-1",
+                "S-2",
                 "성공",
                 new ArticleResponse(article)
         )).orElseGet(() -> RsData.of(
-                "F-1",
-                "%d 번 게시물은 존재하지 않습니다.".formatted(id),
+                "F-1"
+                , "%d 번 게시물은 존재하지 않습니다.".formatted(id),
                 null
         ));
     }
 
+    //게시물 생성 요청 DTO
     @Data
-    public static class WriteRequest {
+    public static class CreateRequest {
         @NotBlank
-        private String title;
+        private String subject;
         @NotBlank
         private String content;
+        @NotBlank
+        private List<Ingredient> ingredients;
     }
 
-    @AllArgsConstructor
+    //게시물 생성 완료 응답 DTO
     @Getter
-    public static class WriteResponse {
-        private final Article article;
+    @AllArgsConstructor
+    public static class CreateResponse {
+        private final Article Article;
     }
 
-
+    //단건 게시물 생성
     @PostMapping("")
-    public RsData<WriteResponse> write(@RequestBody WriteRequest writeRequest) {
-        RsData<Article> writeRs = this.articleService.create(null, writeRequest.getTitle(), writeRequest.getContent());
-        if (writeRs.isFail()) return (RsData) writeRs;
-
-        return RsData.of(
-                writeRs.getResultCode(),
-                writeRs.getMsg(),
-                new WriteResponse(writeRs.getData())
-        );
+    public RsData<CreateResponse> createArticle(@RequestBody CreateRequest CreateRequest) {
+        RsData<Article> createRs = this.articleService.create(null, CreateRequest.getSubject(),
+                CreateRequest.getContent(), CreateRequest.ingredients);
+        if (createRs.isFail()) {
+            return (RsData) createRs;
+        }
+        return RsData.of(createRs.getResultCode(), createRs.getMsg(), new CreateResponse(createRs.getData()));
     }
 
+    //게시물 수정 요청 DTO
     @Data
-    public static class ModifyRequest {
+    public static class UpdateRequest {
         @NotBlank
-        private String title;
+        private String subject;
         @NotBlank
         private String content;
     }
 
-    @AllArgsConstructor
+    //게시물 수정 완료 DTO
     @Getter
-    public static class ModifyResponse {
+    @AllArgsConstructor
+    public static class UpdateResponse {
         private final Article article;
     }
 
+    //단건 게시물 수정
     @PatchMapping("/{id}")
-    public RsData<ModifyResponse> modifyArticle(@PathVariable("id") Long id, @Valid @RequestBody ModifyRequest modifyRequest) {
-        Optional<Article> optionalArticle = this.articleService.findById(id);
-        if (optionalArticle.isEmpty()) {
-            return RsData.of(
-                    "F-4",
-                    "%s번 게시물은 존재하지 않습니다.".formatted(id),
-                    null
-            );
+    public RsData<UpdateResponse> updateArticle(@PathVariable(value = "id") Long id,
+                                                @Valid @RequestBody UpdateRequest updateRequest) {
+        Optional<Article> og = this.articleService.getArticle(id);
+        if (og.isEmpty()) {
+            return RsData.of("F-2", "%d 번 게시물은 존재하지 않습니다.".formatted(id), null);
         }
-        // 회원 권한 체크 canModify();
 
-        RsData<Article> articleRsData = this.articleService.modify(optionalArticle.get(), modifyRequest.getTitle(), modifyRequest.getContent());
-
-        return RsData.of(
-                articleRsData.getResultCode(),
-                articleRsData.getMsg(),
-                new ModifyResponse(articleRsData.getData())
-        );
+        // 수정할 회원의 권한 확인 필요
+        RsData<Article> updateRs = this.articleService.update(og.get(), updateRequest.getSubject(),
+                updateRequest.getContent());
+        return RsData.of(updateRs.getResultCode(), updateRs.getMsg(), new UpdateResponse(updateRs.getData()));
     }
 
-    @AllArgsConstructor
+    //단건 게시물 삭제 응답 DTO
     @Getter
+    @AllArgsConstructor
     public static class DeleteResponse {
         private final Article article;
     }
 
+    //단건 게시물 삭제
     @DeleteMapping("/{id}")
-    public RsData<DeleteResponse> deleteArticle(@PathVariable("id") Long id) {
-        Optional<Article> optionalArticle = this.articleService.findById(id);
-        if (optionalArticle.isEmpty()) {
-            return RsData.of(
-                    "F-5",
-                    "%d번 게시글이 없어 삭제 실패했습니다.".formatted(id),
-                    null
-            );
-        }
-        RsData<Article> deletedRsData = this.articleService.deleteByArticle(optionalArticle.get());
-        return RsData.of(
-                deletedRsData.getResultCode(),
-                deletedRsData.getMsg(),
-                new DeleteResponse(deletedRsData.getData())
-        );
-    }
+    public RsData<DeleteResponse> deleteArticle(@PathVariable(value = "id") Long id) {
 
+        RsData<Article> deleteRs = this.articleService.delete(id);
+
+        return RsData.of(deleteRs.getResultCode(), deleteRs.getMsg(), new DeleteResponse(deleteRs.getData()));
+    }
 }
