+++ src/main/java/com/proj/Matjalal/domain/article/entity/Article.java
@@ -1,14 +1,19 @@
 package com.proj.Matjalal.domain.article.entity;
 
-
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.proj.Matjalal.domain.ingredient.entity.Ingredient;
 import com.proj.Matjalal.domain.member.entity.Member;
 import com.proj.Matjalal.global.jpa.BaseEntity;
 import jakarta.persistence.Entity;
+import jakarta.persistence.ManyToMany;
 import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 @Getter
@@ -16,8 +21,12 @@ import lombok.experimental.SuperBuilder;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Article extends BaseEntity {
-    private String title;
+    private String brand;
+    private String subject;
     private String content;
     @ManyToOne
+    @JsonIgnore
     private Member author;
+    @ManyToMany
+    private List<Ingredient> ingredients = new ArrayList<>();
 }
