+++ src/main/java/com/proj/Matjalal/domain/article/repository/ArticleRepository.java
@@ -1,10 +1,7 @@
 package com.proj.Matjalal.domain.article.repository;
 
-
 import com.proj.Matjalal.domain.article.entity.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-@Repository
 public interface ArticleRepository extends JpaRepository<Article, Long> {
 }
