--- src/main/java/com/proj/Matjalal/domain/gongcha/entity/GongchaArticle.java
@@ -1,31 +1,0 @@
-package com.proj.Matjalal.domain.gongcha.entity;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.proj.Matjalal.domain.ingredient.entity.Ingredient;
-import com.proj.Matjalal.domain.member.entity.Member;
-import com.proj.Matjalal.global.jpa.BaseEntity;
-import jakarta.persistence.Entity;
-import jakarta.persistence.ManyToMany;
-import jakarta.persistence.ManyToOne;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-import org.hibernate.validator.internal.util.stereotypes.Lazy;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@SuperBuilder(toBuilder = true)
-@AllArgsConstructor
-@NoArgsConstructor
-public class GongchaArticle extends BaseEntity {
-    private String subject;
-    private String content;
-    @ManyToOne
-    private Member author;
-    @ManyToMany
-    private List<Ingredient> ingredients = new ArrayList<>();
-}
