--- src/main/java/com/proj/Matjalal/domain/gongcha/repository/GongchaArticleRepository.java
@@ -1,7 +1,0 @@
-package com.proj.Matjalal.domain.gongcha.repository;
-
-import com.proj.Matjalal.domain.gongcha.entity.GongchaArticle;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GongchaArticleRepository extends JpaRepository<GongchaArticle, Long> {
-}
