+++ src/main/java/com/proj/Matjalal/domain/ingredient/entity/Ingredient.java
@@ -1,20 +1,16 @@
 package com.proj.Matjalal.domain.ingredient.entity;
 
-import com.proj.Matjalal.domain.member.entity.Member;
-import com.proj.Matjalal.domain.subway.entitiy.SubwayArticle;
+import com.proj.Matjalal.domain.article.entity.Article;
 import com.proj.Matjalal.global.jpa.BaseEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.ManyToMany;
-import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 @Entity
 @Getter
@@ -26,6 +22,6 @@ public class Ingredient extends BaseEntity {
     private String type;
 
     @ManyToMany(mappedBy = "ingredients")
-    private List<SubwayArticle> subwayArticles = new ArrayList<>();
+    private List<Article> articles = new ArrayList<>();
 }
