+++ src/main/java/com/proj/Matjalal/global/initData/NotProd.java
@@ -24,11 +24,11 @@ public class NotProd {
             Member admin = memberService.join("admin", password, "admin@test.com");
 
             // 작성자 회원 추가
-            articleService.create(user1, "제목 1", "내용 1");
-            articleService.create(user1, "제목 2", "내용 2");
-            articleService.create(user2, "제목 3", "내용 3");
-            articleService.create(user2, "제목 4", "내용 4");
-            articleService.create(admin, "제목 5", "내용 5");
+            articleService.create(user1, "제목 1", "내용 1",null);
+            articleService.create(user1, "제목 2", "내용 2",null);
+            articleService.create(user2, "제목 3", "내용 3",null);
+            articleService.create(user2, "제목 4", "내용 4",null);
+            articleService.create(admin, "제목 5", "내용 5",null);
 
             //재료 생성
             // Subway
