+++ src/main/java/com/proj/Matjalal/global/security/APISecurityConfig.java
@@ -27,15 +27,11 @@ public class APISecurityConfig {
                         authorizeRequests -> authorizeRequests
                                 .requestMatchers("/api/*/articles").permitAll()
                                 .requestMatchers("/api/*/articles/*").permitAll()
-                                .requestMatchers("/api/*/subwayArticles").permitAll()
-                                .requestMatchers("/api/*/subwayArticles/*").permitAll()
                                 .requestMatchers(HttpMethod.POST, "/api/*/members/login").permitAll() // 로그인은 누구나 가능, post 요청만 허용
                                 .requestMatchers(HttpMethod.POST, "/api/*/members/logout").permitAll() // 로그아웃은 누구나 가능
                                 .requestMatchers("/api/*/ingredients").permitAll() //전체 재료 보기는 누구나 가능
                                 .requestMatchers("/api/*/ingredients/*").permitAll() // 재료 상세 보기는 누구나 가능
                                 .requestMatchers("/api/*/ingredients/type/*").permitAll() // 재료 상세 보기는 누구나 가능
-                                .requestMatchers("/api/*/gongchaArticles").permitAll() // 다건요청 권한 허용
-                                .requestMatchers("/api/*/gongchaArticles/*").permitAll() // gongcha 모든 method 권한 허용
                                 .anyRequest().authenticated()
                 )
                 .csrf(
